# frontend/lib/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Atlas Research",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/investigations": {
      "post": {
        "tags": [
          "investigations"
        ],
        "summary": "Create",
        "operationId": "create_api_v1_investigations_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InvestigationCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Investigation"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "503": {
            "description": "Service Unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/investigations/{investigation_id}": {
      "get": {
        "tags": [
          "investigations"
        ],
        "summary": "Investigation",
        "operationId": "investigation_api_v1_investigations__investigation_id__get",
        "parameters": [
          {
            "name": "investigation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Investigation Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Investigation"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/investigations/{investigation_id}/events": {
      "get": {
        "tags": [
          "investigations"
        ],
        "summary": "Events",
        "operationId": "events_api_v1_investigations__investigation_id__events_get",
        "parameters": [
          {
            "name": "investigation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Investigation Id"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 500,
              "minimum": 1,
              "description": "page size",
              "default": 50,
              "title": "Limit"
            },
            "description": "page size"
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "description": "items to skip",
              "default": 0,
              "title": "Offset"
            },
            "description": "items to skip"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_InvestigationEvent_"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/investigations/{investigation_id}/resume": {
      "post": {
        "tags": [
          "investigations"
        ],
        "summary": "Resume",
        "operationId": "resume_api_v1_investigations__investigation_id__resume_post",
        "parameters": [
          {
            "name": "investigation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Investigation Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResumeRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Investigation"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        }
      }
    },
    "/api/v1/investigations/{investigation_id}/premises/{premise_key}/disprove": {
      "post": {
        "tags": [
          "investigations"
        ],
        "summary": "Disprove",
        "operationId": "disprove_api_v1_investigations__investigation_id__premises__premise_key__disprove_post",
        "parameters": [
          {
            "name": "investigation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Investigation Id"
            }
          },
          {
            "name": "premise_key",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Premise Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DisproveRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Investigation"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "BudgetRequest": {
        "properties": {
          "max_rounds": {
            "type": "integer",
            "maximum": 2.0,
            "minimum": 1.0,
            "title": "Max Rounds",
            "default": 2
          },
          "max_leads": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 10.0,
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Leads"
          },
          "max_documents": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 25.0,
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Documents"
          },
          "token_budget": {
            "anyOf": [
              {
                "type": "integer",
                "exclusiveMinimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Token Budget"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "BudgetRequest",
        "description": "Per-run budgets (spec §7.4); unset ones take the configured defaults."
      },
      "Budgets": {
        "properties": {
          "max_rounds": {
            "type": "integer",
            "title": "Max Rounds"
          },
          "max_leads": {
            "type": "integer",
            "title": "Max Leads"
          },
          "max_documents": {
            "type": "integer",
            "title": "Max Documents"
          },
          "token_budget": {
            "type": "integer",
            "title": "Token Budget"
          }
        },
        "type": "object",
        "required": [
          "max_rounds",
          "max_leads",
          "max_documents",
          "token_budget"
        ],
        "title": "Budgets",
        "description": "Per-run limits (spec §7.4)."
      },
...
      "DisproveRequest": {
        "properties": {
          "reason": {
            "type": "string",
            "maxLength": 2000,
            "minLength": 1,
            "pattern": "\\S",
            "title": "Reason"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "reason"
        ],
        "title": "DisproveRequest"
      },
      "ErrorDetail": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorDetail"
      },
      "ErrorEnvelope": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          }
        },
        "type": "object",
        "required": [
          "error"
        ],
        "title": "ErrorEnvelope"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "Investigation": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "theme": {
            "type": "string",
            "title": "Theme"
          },
          "question": {
            "type": "string",
            "title": "Question"
          },
          "status": {
            "type": "string",
            "enum": [
              "running",
              "stopped"
            ],
            "title": "Status"
          },
          "stop_reason": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "answered",
                  "no_new_independent_evidence",
                  "budget_exhausted",
                  "needs_review",
                  "premise_disproven"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Stop Reason"
          },
          "stop_detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Stop Detail"
          },
          "paused": {
            "type": "boolean",
            "title": "Paused"
          },
          "resumable": {
            "type": "boolean",
            "title": "Resumable"
          },
          "run_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Run Id"
          },
          "request": {
            "$ref": "#/components/schemas/InvestigationRequest"
          },
          "budgets": {
            "$ref": "#/components/schemas/Budgets"
          },
          "usage": {
            "$ref": "#/components/schemas/Usage"
          },
          "premises": {
            "items": {
              "$ref": "#/components/schemas/Premise"
            },
            "type": "array",
            "title": "Premises"
          },
          "tasks": {
            "items": {
              "$ref": "#/components/schemas/Task"
            },
            "type": "array",
            "title": "Tasks"
          },
          "leads": {
            "items": {
              "$ref": "#/components/schemas/InvestigationLead"
            },
            "type": "array",
            "title": "Leads"
          },
          "documents": {
            "items": {
              "$ref": "#/components/schemas/InvestigationDocument"
            },
            "type": "array",
            "title": "Documents"
          },
          "research_card": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ResearchCard"
              },
              {
                "type": "null"
              }
            ]
          },
          "created_by": {
            "type": "string",
            "title": "Created By"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "stopped_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Stopped At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "theme",
          "question",
          "status",
          "stop_reason",
          "stop_detail",
          "paused",
          "resumable",
          "run_id",
          "request",
          "budgets",
          "usage",
          "premises",
          "tasks",
          "leads",
          "documents",
          "research_card",
          "created_by",
          "created_at",
          "stopped_at"
        ],
        "title": "Investigation"
      },
      "InvestigationCreate": {
        "properties": {
          "theme": {
            "type": "string",
            "minLength": 1,
            "title": "Theme",
            "description": "a theme in the universe config"
          },
          "question": {
            "type": "string",
            "maxLength": 4000,
            "minLength": 1,
            "pattern": "\\S",
            "title": "Question",
            "description": "the research question"
          },
          "seed_company_ids": {
            "anyOf": [
              {
                "items": {
                  "type": "string",
                  "format": "uuid"
                },
                "type": "array",
                "maxItems": 25,
                "minItems": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Seed Company Ids",
            "description": "the companies to investigate; default: the theme's companies"
          },
          "as_of": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "As Of",
            "description": "only Source Versions available by then are read; default now"
          },
          "budgets": {
            "$ref": "#/components/schemas/BudgetRequest"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "theme",
          "question"
        ],
        "title": "InvestigationCreate"
      },
      "InvestigationDocument": {
        "properties": {
          "source_version_id": {
            "type": "string",
            "format": "uuid",
            "title": "Source Version Id"
          },
          "task_key": {
            "type": "string",
            "title": "Task Key"
          },
          "company_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Company Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "available_at": {
            "type": "string",
            "format": "date-time",
            "title": "Available At"
          }
        },
        "type": "object",
        "required": [
          "source_version_id",
          "task_key",
          "company_id",
          "title",
          "available_at"
        ],
        "title": "InvestigationDocument",
        "description": "A Source Version an Investigator task read (counted against `max_documents`)."
      },
...
      "InvestigationLead": {
        "properties": {
          "rank": {
            "type": "integer",
            "title": "Rank"
          },
          "lead_id": {
            "type": "string",
            "format": "uuid",
            "title": "Lead Id"
          },
          "discovery_id": {
            "type": "string",
            "format": "uuid",
            "title": "Discovery Id"
          },
          "url": {
            "type": "string",
            "title": "Url"
          },
          "canonical_url": {
            "type": "string",
            "title": "Canonical Url"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "snippet": {
            "type": "string",
            "title": "Snippet"
          },
          "tier": {
            "type": "string",
            "const": "C",
            "title": "Tier"
          }
        },
        "type": "object",
        "required": [
          "rank",
          "lead_id",
          "discovery_id",
          "url",
          "canonical_url",
          "title",
          "snippet",
          "tier"
        ],
        "title": "InvestigationLead",
        "description": "A Tier C lead the Scout found and the investigation kept (never Evidence)."
      },
      "InvestigationRequest": {
        "properties": {
          "run_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Run Id"
          },
          "research_question": {
            "type": "string",
            "title": "Research Question"
          },
          "theme_id": {
            "type": "string",
            "title": "Theme Id"
          },
          "seed_entity_ids": {
            "items": {
              "type": "string",
              "format": "uuid"
            },
            "type": "array",
            "title": "Seed Entity Ids"
          },
          "as_of_utc": {
            "type": "string",
            "format": "date-time",
            "title": "As Of Utc"
          },
          "allowed_source_tiers": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Allowed Source Tiers"
          },
          "available_budget": {
            "$ref": "#/components/schemas/Budgets"
          },
          "max_depth": {
            "type": "integer",
            "title": "Max Depth"
          },
          "max_new_leads": {
            "type": "integer",
            "title": "Max New Leads"
          },
          "approved_tool_list": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Approved Tool List"
          },
          "relevant_hindsight_bank": {
            "type": "string",
            "title": "Relevant Hindsight Bank"
          },
          "hypothesis_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Hypothesis Id"
          }
        },
        "type": "object",
        "required": [
          "run_id",
          "research_question",
          "theme_id",
          "seed_entity_ids",
          "as_of_utc",
          "allowed_source_tiers",
          "available_budget",
          "max_depth",
          "max_new_leads",
          "approved_tool_list",
          "relevant_hindsight_bank",
          "hypothesis_id"
        ],
        "title": "InvestigationRequest",
        "description": "The shared agent contract's request (spec §7.2), as every role of the run works to."
      },
...
      "Page_InvestigationEvent_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/InvestigationEvent"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "limit": {
            "type": "integer",
            "title": "Limit"
          },
          "offset": {
            "type": "integer",
            "title": "Offset"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "limit",
          "offset"
        ],
        "title": "Page[InvestigationEvent]"
      },
...
      "Premise": {
        "properties": {
          "key": {
            "type": "string",
            "title": "Key"
          },
          "statement": {
            "type": "string",
            "title": "Statement"
          },
          "company_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Company Id"
          },
          "status": {
            "type": "string",
            "enum": [
              "open",
              "disproven"
            ],
            "title": "Status"
          },
          "reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          },
          "disproven_by": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Disproven By"
          },
          "disproven_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Disproven At"
          }
        },
        "type": "object",
        "required": [
          "key",
          "statement",
          "company_id",
          "status",
          "reason",
          "disproven_by",
          "disproven_at"
        ],
        "title": "Premise"
      },
...
      "ResumeRequest": {
        "properties": {
          "token_budget": {
            "type": "integer",
            "exclusiveMinimum": 0.0,
            "title": "Token Budget",
            "description": "the new token budget; larger than the old one"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "token_budget"
        ],
        "title": "ResumeRequest"
      },
...
      "Usage": {
        "properties": {
          "rounds": {
            "type": "integer",
            "title": "Rounds"
          },
          "leads": {
            "type": "integer",
            "title": "Leads"
          },
          "documents": {
            "type": "integer",
            "title": "Documents"
          },
          "tokens_in": {
            "type": "integer",
            "title": "Tokens In"
          },
          "tokens_out": {
            "type": "integer",
            "title": "Tokens Out"
          }
        },
        "type": "object",
        "required": [
          "rounds",
          "leads",
          "documents",
          "tokens_in",
          "tokens_out"
        ],
        "title": "Usage",
        "description": "How much of each budget the investigation has used."
      },
...
    }
  }
}